requirement.txt
streamlit 
langchain
openai
requests
chromadb
frontend
fitz
langchain-openai
langchain-community
pymupdf 
pillow
pypdf